bash, dir.xml parsing, unix
The purpose of this assignment is for students to become comfortable with basic commands, documentation, file redirection - i.e. the bash shell), a Unix editor, the Unix file system using Unix filters (including Awk) to solve text processing problems.

Your task is to recurse through a subtree of the filesystem and put a file called dir.xml in each directory. You will pull data from 2 sources: the directory itself, and a file that may appear in each directory called README .